Major patch:_Redesigned pipe model and backends
The previous design placed the responsibility of thread safety
and synchronization, including the callback when data is ready
on a pipe, to the backend implementing the pipe, like DLCI
channels and the UART backend.

This resulted in a bunch of repeated code and instability since
there was no contract synchronizing changing the callback and
user data stored in the pipe struct.

The new design makes the modem_pipe responsible for thread
safety and managing the callback and raising events. It
now behaves more like a linux pipe.

It can be:
Opened, which will raise an event and unblocks api calls
Attached, which will set the callback and user data
Released, which clears the callback and user data
Closed, which raised an event, and blocks api calls

Furthermore, the modem_pipe_api now has api calls for
open and close, which are invoked when a user wants to
open or close the pipe using the modem_pipe_open and
modem_pipe_close functions. This allows for synchronizing
the backend with the pipe, flushing the uart and enabling
rx/tx when opened, and disabling rx/tx when closed.

There are also two added functions for synchronously
opening and closing a pipe, which prevents a bunch of
boiler plate code when waiting for a pipe to open/close.

Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>