Feature: Added backend for POSIX TTY pipe
A backend for tty pipes has been added, and a sample application
using it has been added cmux_ppp_tty. It has been tested to work
with the Ublox SARA-R412 using a FTDI RS232 to TTL converter.

A bug fix for the uart backend has also been added, which delegates
the receive ready callback to the work queue.

Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>