modem_cmux: Patched disconnect and dlci open/close bugs
The closedown ack was ignored by mistake, leaving the modem_cmux
module resending closedown command until pipe was released. This
has been patched, and unit tests have been added which catches
the bug.

When opening a DLCI channel, a bug caused the DLCI struct to
be added to the single linked list which contains all DLCIs
instances bound to the CMUX instance, every time it was opened.
The DLCI instance should only be added once, during the init.
This bug has been patched, and is covered by the unit test
added to fix the previous bug.

Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>