patch: Updated backend mock API
The backend mock can now be used to test syncronous
transactions:

1. mock sets up expected data and response to said data
2. module sends data
3. mock validates sent data is expected
4. mock responds with set up response
5. module gets response
6. test continues

This allows for testing sync functions like
modem_pipe_open, within which a bug was found, that is
now recreatable in the unit testing.

The bug caused modem_pipe_open and modem_pipe_close to
not wait until closed. This bug has been fixed as part
of this commit.

Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>