modem_cmux: Added FCon and FCoff support
The modem_cmux module now supports software
flow control, using the control channel commands
FCon and FCoff. It does not support per DLCI flow
control.

With this change, an update to the frame transmision
has been added, which splits the transmit_frame() into
transmit_cmd_frame() and transmit_data_frame().

transmit_cmd_frame() will send the frame if there is
space for it on the transmit buffer, and is used for
the control channel, DLCI0. To ensure it can always
send control frames, space is allocated on the
transmit buffer for two control frames.

transmit_data_frame() will only add frames to the
transmit buffer if there is room for at least 8
bytes of data, alongside the wrapping for the
dataframe, and there is room for the two reserved
command frames.

This ensures that flow control, which disables
transmission of data frames only, will not block
control frames from being added to the transmit
buffer while flow control is active, which is
essential to acknowledge the FCon and FCoff
frames for example.

Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>