CMUX API Update: default functions are now synchronous
The modem_cmux_connect and modem_cmux_disconnect
functions are now synchronous. This change makes all
modules follow the same naming convention and behavior.

The default behavior of any call which does not have
_async appended to the function name is blocking,
until the action has been performed. This allows
the user to choose if the modules should be used
asynchronously or not, without requiring the user
to implement their own external synchronization.

This is a breaking change, since the behavior of
the functions have changed.

Tests have been added to cover the synchronous
use of the functions. And a script has been added
to build and run all tests automatically for
convenience. For the end implementation, we will
use twister.

Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>