Feature: Add UART async API backend
This release cleans up the backends, moving them into
their own subfolder. The UART backend has been updated
to support both interrupt driven and async (DMA) APIs.

The interface for the UART backend has not changed. The
backend automatically chooses the appropriate API based
on the capabilities of the UART. If both interrupt
driven and async APIs are enabled using Kconfig, the
backend will prefer the async API, and fallback to the
interrupt driven API if async API is not supported or
configured for the targeted UART device.

Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>