Feature: Added generic driver
The driver is meant to be used as a template to
implement proper drivers for modems. It currently
works naively, with little error handling.

The driver is async, and will configure the modem
and bring up the network in the background without
delaying bootup.

The driver will try to re register and set up PPP
if it looses the network.

Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>